perf: move form input watcher into PublicFormSubmitButton subcomponent
that component is split up so that input changes will not rerender the entire FormFields component leading to terrible performance.
